separated add and edit notes into components
