+++ src/components/ApplicationView.js
@@ -6,7 +6,7 @@ import { MedicalRoutes } from "./routes/MedicalRoutes";
 import { NoteRoutes } from "./routes/NoteRoutes";
 import { PetRoutes } from "./routes/PetRoutes";
 
-export const ApplicationView = ({user, updateUser, pet, myPets, setMyPets}) => {
+export const ApplicationView = ({user, updateUser, pet, myPets, setMyPets, defaultPet}) => {
     return (
         <>
             <Route exact path="/">
@@ -15,7 +15,7 @@ export const ApplicationView = ({user, updateUser, pet, myPets, setMyPets}) => {
             <PetRoutes user={user} updateUser={updateUser} pet={pet} myPets={myPets} setMyPets={setMyPets} />
             <MedicalRoutes />
             <ContactRoutes />
-            <NoteRoutes />
+            <NoteRoutes user={user} updateUser={updateUser} defaultPet={defaultPet} />
         </>
     )
 }
