+++ src/components/notes/Note.js
@@ -15,40 +15,23 @@ import TextField from "@mui/material/TextField";
 import StarBorderIcon from "@mui/icons-material/StarBorder";
 import StarIcon from "@mui/icons-material/Star";
 import IconButton from "@mui/material/IconButton";
+import { EditNoteDialog } from "./EditNoteDialog";
 
 export const Note = ({ note, syncNotes }) => {
-    const [editedNote, updateEditedNote] = useState(note);
     const [open, setOpen] = useState(false);
-
-    useEffect(() => {
-        updateEditedNote(note)
-    }, [note])
-
+    
     const deleteNote = () => [
         NotesRepository.delete(note.id).then(() => syncNotes()),
     ];
-    const noteDateTime = new Date(note.timestamp).toLocaleString();
 
-    const editNote = (event) => {
-        event.preventDefault();
-        handleClose();
-        const copy = { ...editedNote };
-        copy.timestamp = Date.now();
-        NotesRepository.editNote(copy).then(() => syncNotes());
-    };
-
-    const starUnstar = (event) => {
-        const copy = { ...editedNote };
-        copy.starred = !editedNote.starred;
+    const starUnstar = () => {
+        const copy = { ...note };
+        copy.starred = !note.starred;
         NotesRepository.editNote(copy).then(() => syncNotes());
     };
 
     const handleClickOpen = () => {
         setOpen(true);
-    };
-
-    const handleClose = () => {
-        setOpen(false);
     };
 
     return (
@@ -73,44 +56,7 @@ export const Note = ({ note, syncNotes }) => {
                     {note.starred ? <StarIcon /> : <StarBorderIcon />}
                 </IconButton>
             </CardActions>
-            <Dialog open={open} onClose={handleClose}>
-                <form onSubmit={editNote}>
-                    <DialogTitle>Add Note</DialogTitle>
-                    <DialogContent>
-                        <TextField
-                            margin="dense"
-                            id="subject"
-                            label="subject"
-                            required
-                            value={editedNote.subject}
-                            type="text"
-                            onChange={(event) => {
-                                const copy = { ...editedNote };
-                                copy.subject = event.target.value;
-                                updateEditedNote(copy);
-                            }}
-                        />
-                        <TextField
-                            margin="dense"
-                            id="textBody"
-                            label="textBody"
-                            required
-                            value={editedNote.textBody}
-                            multiline
-                            type="text"
-                            onChange={(event) => {
-                                const copy = { ...editedNote };
-                                copy.textBody = event.target.value;
-                                updateEditedNote(copy);
-                            }}
-                        />
-                    </DialogContent>
-                    <DialogActions>
-                        <Button onClick={handleClose}>Cancel</Button>
-                        <Button type="submit">Save</Button>
-                    </DialogActions>
-                </form>
-            </Dialog>
+            <EditNoteDialog open={open} setOpen={setOpen} syncNotes={syncNotes} note={note}/>
         </Card>
     );
 };
