+++ src/components/notes/NotesList.js
@@ -1,101 +1,31 @@
 import React, { useEffect, useState } from "react";
-import { useHistory } from "react-router-dom";
 import useSimpleAuth from "../../hooks/ui/useSimpleAuth";
 import { Note } from "./Note";
-import PetRepository from "../../repositories/PetRepository";
-import UserRepository from "../../repositories/UserRepository";
 import NotesRepository from "../../repositories/NotesRepository";
 
 import Container from "@mui/material/Container";
 import Grid from "@mui/material/Grid";
 import Typography from "@mui/material/Typography";
 import Button from "@mui/material/Button";
-import Dialog from "@mui/material/Dialog";
-import DialogActions from "@mui/material/DialogActions";
-import DialogContent from "@mui/material/DialogContent";
-import DialogTitle from "@mui/material/DialogTitle";
-import TextField from "@mui/material/TextField";
+import { AddNoteDialog } from "./AddNoteDialog";
 
-export const NotesList = () => {
+export const NotesList = ({ user, updateUser, defaultPet }) => {
     const [myPetNotes, setMyPetNotes] = useState([]);
-    const [newNote, setNewNote] = useState({})
-    const [user, updateUser] = useState({});
-    const [myPets, setMyPets] = useState([]);
-    const [defaultPet, setDefaultPet] = useState({});
-    const { getCurrentUser } = useSimpleAuth();
-    const history = useHistory();
     const [open, setOpen] = useState(false);
 
-    const syncPets = () => {
-        PetRepository.getAllExpandAllByUser(getCurrentUser().id).then(
-            (data) => {
-                data.sort((el1) => {
-                    if (el1.id === user.defaultPetId) {
-                        return -1;
-                    }
-                });
-                setMyPets(data);
-            }
+    const syncNotes = () => {
+        NotesRepository.findNotesByPet(defaultPet?.id).then((data) =>
+            setMyPetNotes(data)
         );
     };
-
-    const syncNotes = () => {
-        NotesRepository.findNotesByPet(user.defaultPetId)
-            .then(data => setMyPetNotes(data))
-    }
-
-    const syncUser = () => {
-        UserRepository.get(getCurrentUser().id).then((data) => {
-            updateUser(data);
-        });
-    };
-
-    useEffect(() => {
-        syncUser();
-    }, []);
-
-    useEffect(() => {
-        syncPets();
-        syncNotes();
-    }, [user]);
-
-    useEffect(() => {
-        setDefaultPet(myPets.find((pet) => user.defaultPetId === pet.id));
-        return () => {
-            setDefaultPet({});
-        };
-    }, [myPets]);
-
-    useEffect(() => {
-        
-    }, [defaultPet])
-
-    useEffect(() => {
-        syncUser();
-        return () => {
-            syncUser({});
-        };
-    }, [history.location.state]);
-
-    const addNote = (event) => {
-        event.preventDefault();
-        handleClose();
-        const copy = {...newNote};
-        copy.starred = false;
-        copy.userId = user.id;
-        copy.petId = user.defaultPetId;
-        copy.timestamp = Date.now();
-        NotesRepository.addNote(copy)
-            .then(() => syncNotes())
-    }
 
     const handleClickOpen = () => {
         setOpen(true);
     };
 
-    const handleClose = () => {
-        setOpen(false);
-    };
+    useEffect( () => {
+        syncNotes();
+    }, [defaultPet])
 
     return (
         <Container maxWidth="lg">
@@ -114,46 +44,16 @@ export const NotesList = () => {
             <Grid container spacing={4} sx={{ justifyContent: "center" }}>
                 {myPetNotes.map((note) => (
                     <Grid item sm={4} lg={3} key={`note--${note.id}`}>
-                        <Note note={note} syncNotes={syncNotes} key={note.id}/>
+                        <Note note={note} syncNotes={syncNotes} key={note.id} />
                     </Grid>
                 ))}
             </Grid>
-            <Dialog open={open} onClose={handleClose}>
-                <form onSubmit={addNote}>
-                    <DialogTitle>Add Note</DialogTitle>
-                    <DialogContent>
-                        <TextField
-                            margin="dense"
-                            id="subject"
-                            label="subject"
-                            required
-                            type="text"
-                            onChange={(event) => {
-                                const copy = { ...newNote };
-                                copy.subject = event.target.value;
-                                setNewNote(copy);
-                            }}
-                        />
-                        <TextField
-                            margin="dense"
-                            id="textBody"
-                            label="textBody"
-                            required
-                            multiline
-                            type="text"
-                            onChange={(event) => {
-                                const copy = { ...newNote };
-                                copy.textBody = event.target.value;
-                                setNewNote(copy);
-                            }}
-                        />
-                    </DialogContent>
-                    <DialogActions>
-                        <Button onClick={handleClose}>Cancel</Button>
-                        <Button type="submit">Save</Button>
-                    </DialogActions>
-                </form>
-            </Dialog>
+            <AddNoteDialog
+                open={open}
+                setOpen={setOpen}
+                user={user}
+                syncNotes={syncNotes}
+            />
         </Container>
     );
 };
