+++ src/components/routes/NoteRoutes.js
@@ -2,11 +2,11 @@ import React from "react";
 import { Route } from "react-router-dom";
 import { NotesList } from "../notes/NotesList";
 
-export const NoteRoutes = () => {
+export const NoteRoutes = ({user, updateUser, defaultPet}) => {
     return (
         <>
             <Route exact path="/notes">
-                <NotesList />
+                <NotesList user={user} updateUser={updateUser} defaultPet={defaultPet} />
             </Route>
         </>
     )
